removes duplcated test
